Move download gerber modal to views
