+++ js/view.js
@@ -4,7 +4,6 @@ import { html } from "lit-html";
 import "./codemirror/codemirror.js";
 import { files } from "./components-names.js";
 import { downloadSVG, downloadText, downloadPNG } from "./events/download.js"
-import { downloadGerber } from "./events/downloadGerber.js"
 import { drawImportItems } from "./views/drawImportItems.js";
 import { drawComponentMenu } from "./views/drawComponentMenu.js";
 import { layersColorPicker } from "./views/layersColorPicker.js";
@@ -15,6 +14,7 @@ import { clearSelectedPath } from "./clearSelectedPath.js";
 import logoURL from '../logo.svg'
 import { inputRenderers } from "./views/inputRenderers.js";
 import { initCodeMirror } from "./codemirror/codemirror.js";
+import { drawDownloadGerberModal } from "./views/drawDownloadGerberModal.js";
 
 export function view(state) {
 	return html`
@@ -54,7 +54,7 @@ export function view(state) {
 			${drawComponentMenu(files)}
 		</div>
 
-		${modal_download_gerber(state)}
+		${drawDownloadGerberModal(state)}
 	`
 }
 
@@ -85,9 +85,7 @@ const menu = state => html`
 					</div class="menu-item">
 					<div class="menu-item"
 						@click=${(e) => {
-							setDownloadGerberLayerVisibility(state);
-							const modal = document.getElementById("modal_download_gerber");
-							modal.classList.remove("hidden");
+							state.downloadGerberModal = true;
 						}}>
 						gerber (WIP)
 					</div class="menu-item">
@@ -196,136 +194,3 @@ const menu = state => html`
 		</a>
 	</div>
 `;
-
-function getLayerList(layerMap){
-	const list = [];
-	layerMap.forEach((val, key) => {
-		list.push(key);
-	});
-	return list;
-}
-
-function setDownloadGerberLayerVisibility(state){
-	// Set layer visiblity based on what layers exist
-	state.downloadGerberOptions.layers.set("F.Cu", state.pcb.layers["F.Cu"] === undefined ? false : true);
-	state.downloadGerberOptions.layers.set("B.Cu", state.pcb.layers["B.Cu"] === undefined ? false : true);
-	state.downloadGerberOptions.layers.set("F.Mask", state.pcb.layers["F.Cu"] === undefined ? false : true);
-	state.downloadGerberOptions.layers.set("B.Mask", state.pcb.layers["B.Cu"] === undefined ? false : true);
-	state.downloadGerberOptions.layers.set("Outline", state.pcb.layers["interior"] === undefined ? false : true);
-	state.downloadGerberOptions.layers.set("Drills", state.pcb.layers["drill"] === undefined ? false : true);
-}
-
-const modal_download_gerber = state => html`
-<div id="modal_download_gerber" class="modal hidden">
-	<div class="modal-content">
-		<div class="modal-header">
-			<div class="col-75 align-left">
-				<i class="icon fa fa-arrow-circle-down"></i>
-				<h3 class="modal-title">Download Gerber Options</h3>
-			</div>
-			<div class="col-25 align-right">
-				<span 
-					class="close"
-					@click=${(e) => {
-						const modal = document.getElementById("modal_download_gerber");
-						modal.classList.add("hidden");
-					}}><i class="fa fa-times"></i></span>
-			</div>
-		</div>
-		<div class="modal-body">
-			<div class="col-50">
-				<h4>Include Layers</h4>
-				${
-					getLayerList(state.downloadGerberOptions.layers).map((l) => {
-					return html`
-					<div class="modal-line">
-						<input 
-							type="checkbox"
-							id="input-${l}"
-							.checked=${state.downloadGerberOptions.layers.get(l)}
-							@change=${(e) => {
-								state.downloadGerberOptions.layers.set(l, e.target.checked);
-							}}> 
-						<label for="input-${l}">${l}</label>
-					</div>
-					`
-				})}
-			</div> <!-- /.col-50 -->
-			<div class="col-50">
-				<h4><label for="projectName">Project Name</label></h4>
-				<div class="modal-line">
-					<input id="projectName" 
-						type="text" 
-						.value=${state.name === "" ? "Untitled" : state.name}
-						@input=${(e) => {
-							state.name = e.target.value;
-						}}>
-				</div>
-
-				<h4>Excellon Drill Units</h4>
-				<div class="modal-line">	
-					<input 
-						type="radio"
-						id="input-excellonMM"
-						name="excellonUnits"
-						.checked=${state.downloadGerberOptions.excellonMetric}
-						@change=${(e) => {
-							state.downloadGerberOptions.excellonMetric = e.target.checked;
-						}}>
-					<label for="input-excellonMM">mm</label>
-					&nbsp;
-					<input 
-						type="radio"
-						id="input-excellonIN"
-						name="excellonUnits"
-						.checked=${!state.downloadGerberOptions.excellonMetric}
-						@change=${(e) => {
-							state.downloadGerberOptions.excellonMetric = !e.target.checked;
-						}}>
-					<label for="input-excellonIN">in</label>
-				</div> <!-- /.modal-line -->
-
-				<h4>Additional Options</h4>
-				<div class="modal-line">
-					<input 
-						type="checkbox"
-						id="input-outlineAllLayers"
-						.checked=${state.downloadGerberOptions.includeOutline}
-						@change=${(e) => {
-							state.downloadGerberOptions.includeOutline = e.target.checked;
-						}}> 
-					<label for="input-outlineAllLayers">Inclue outline in all layers</label>
-				</div>
-				<div class="modal-line">
-					<input 
-						type="checkbox"
-						id="input-protelFilenames"
-						.checked=${state.downloadGerberOptions.protelFilenames}
-						@change=${(e) => {
-							state.downloadGerberOptions.protelFilenames = e.target.checked;
-						}}> 
-					<label for="input-protelFilenames">Use Protel Filenames</label>
-				</div>
-			</div> <!-- /.col-50 -->
-		</div> <!-- /.modal-body -->
-		<div class="modal-footer">
-			<button 
-				type="button" 
-				class="btn"
-				@click=${(e) => {
-					const modal = document.getElementById("modal_download_gerber");
-					modal.classList.add("hidden");
-				}}>Cancel</button>
-			<button 
-				type="button" 
-				class="btn btn-primary"
-				@click=${(e) => {
-					const modal = document.getElementById("modal_download_gerber");
-					modal.classList.add("hidden");
-					downloadGerber(state);
-				}}>Download</button>
-		</div>
-	</div>
-</div>
-`;
-
