+++ styles.css
@@ -642,10 +642,6 @@ input {
 	margin: 7px 0 7px 0;
 }
 
-.hidden {
-	display: none;
-}
-
 .btn {
 	height: 30px;
 	padding: 7px;
